refactor: little refactoring for some core files
